Stop Validation on Ignored NPM Packages
Fixes https://github.com/github/licensed/issues/98

An NPM package could be ignored, but the NPM source would fail if it was
unable to find the package within the node_modules directory. If we are
ignoring the package, then we should not be concerned with that
package's existence in node_modules.

Skip over the assets in building the dependencies list if the package is
ignored.