Handle plugins block and empty url
Most more advanced build.gradle files out in the wild has a plugins block,
buildscript block or both. At the moment there is a restriction
where plugins blocks can be added, so the current implementation
cant handle the "apply build.gradle" clause when generating the
tempfile.

This fix will check if the "Gradle license report" plugin is installed,
and if not, either add it to the plugins block, or add an
plugins block with it if there was none in the build.gradle file.
This should be compatible with all the cases first mentioned.

Also, there was a problem where if one dependency was missing an url,
the check would fail with exception.
This takes an easy way out to fix this - it sets it to an empty
mutable string.